+++ packages/firebase_storage_util/example/lib/main.dart
@@ -1,10 +1,13 @@
+import 'dart:io';
+
 import 'package:flutter/material.dart';
+import 'package:firebase_storage_util/firebase_storage_util.dart';
 
 void main() {
-  runApp(MyApp());
+  runApp(FirebaseStorageUtilExample());
 }
 
-class MyApp extends StatelessWidget {
+class FirebaseStorageUtilExample extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -18,8 +21,20 @@ class MyHomePage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      color: Colors.red,
+    return Scaffold(
+      body: Center(
+        child: ElevatedButton(
+          onPressed: () async {
+            String? url = await FirebaseStorageUtil.instance.upload(
+              uploadPath: 'example/path',
+              uploadDataSource: UploadDataSource.file,
+              file: File('path/to/file/on/device'),
+              getDownloadUrl: true,
+            );
+          },
+          child: Text('Upload'),
+        ),
+      ),
     );
   }
 }