+++ packages/firebase_storage_util/CHANGELOG.md
@@ -3,9 +3,22 @@
 # Changelog 
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+<!-- ## [Unreleased] -->
+
+
+## [0.0.2] - 2021-07-10
 ### Added
 - `deleteFromPath()`, `deleteFromUrl()`, `delete()` methods. 
+- Added extra documentation 
+- Added to the example folder and added example to the readme
+
+### Changed
+- `uploadFromPath()`, `uploadFromFile()` and `uploadFromBytes()` are now public,
+  usable alone without `upload()`, and can potentilally return a downloadUrl.
+- Changed class name from `CustomStorageUtil` to `FirebaseStorageUtil` to match
+  library name and be more representative of this library's function.
+
+
 
 
 ## [0.0.1] - 2021-07-10